admin_panel/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9.9
django-unfold>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
